move model config, server and op back to op folder
