+++ uniflow/flow/transform/transform_huggingface_flow.py
@@ -3,10 +3,9 @@ from typing import Any, Dict, Sequence
 
 from uniflow.constants import TRANSFORM
 from uniflow.flow.flow import Flow
-from uniflow.flow.transform.model import LLMDataProcessor
 from uniflow.node.node import Node
 from uniflow.op.prompt_schema import GuidedPrompt
-from uniflow.op.transform.model_op import ModelOp
+from uniflow.op.transform.model_op import LLMDataProcessor, ModelOp
 
 
 class HuggingFaceModelFlow(Flow):
