+++ uniflow/flow/transform/transform_lmqg_flow.py
@@ -2,10 +2,9 @@ from typing import Any, Dict, Sequence
 
 from uniflow.constants import TRANSFORM
 from uniflow.flow.flow import Flow
-from uniflow.flow.transform.model import LLMDataProcessor
 from uniflow.node.node import Node
 from uniflow.op.prompt_schema import GuidedPrompt
-from uniflow.op.transform.model_op import ModelOp
+from uniflow.op.transform.model_op import LLMDataProcessor, ModelOp
 
 
 class TransformLMQGFlow(Flow):
