+++ uniflow/flow/transform/transform_openai_flow.py
@@ -3,10 +3,10 @@ from typing import Any, Dict, Sequence
 
 from uniflow.constants import TRANSFORM
 from uniflow.flow.flow import Flow
-from uniflow.flow.transform.model import LLMDataProcessor, LLMDataProcessorJson
+from uniflow.flow.transform.model import LLMDataProcessorJson
 from uniflow.node.node import Node
 from uniflow.op.prompt_schema import GuidedPrompt
-from uniflow.op.transform.model_op import ModelOp
+from uniflow.op.transform.model_op import LLMDataProcessor, ModelOp
 
 
 class OpenAIModelFlow(Flow):
